Refactor AI command to streamline diff formatting and remove unused code
