+++ src-tauri/src/commands/remotes.rs
@@ -2,58 +2,8 @@ use crate::error::{AxisError, Result};
 use crate::events::{GitOperationType, ProgressStage};
 use crate::models::{FetchOptions, FetchResult, PullOptions, PushOptions, PushResult, Remote};
 use crate::services::ProgressContext;
-use crate::state::{AppState, GitServiceHandle};
+use crate::state::AppState;
 use tauri::State;
-
-/// Build the refs stdin string for the pre-push hook
-/// Format: <local ref> <local sha> <remote ref> <remote sha>\n per ref
-async fn build_push_refs_stdin(
-    git_service: &GitServiceHandle,
-    remote_name: &str,
-    refspecs: &[String],
-) -> String {
-    let mut refs_lines = Vec::new();
-
-    for refspec in refspecs {
-        // Parse refspec (e.g., "refs/heads/main:refs/heads/main" or just "main")
-        let (local_ref, remote_ref) = if refspec.contains(':') {
-            let parts: Vec<&str> = refspec.split(':').collect();
-            (parts[0].to_string(), parts[1].to_string())
-        } else {
-            // Simple branch name
-            let full_ref = format!("refs/heads/{refspec}");
-            (full_ref.clone(), full_ref)
-        };
-
-        // Get local SHA
-        let local_sha = git_service
-            .read()
-            .await
-            .resolve_ref(&local_ref)
-            .await
-            .unwrap_or_else(|| "0".repeat(40));
-
-        // Get remote SHA (what the remote currently has)
-        let remote_ref_name = format!(
-            "refs/remotes/{remote_name}/{}",
-            refspec
-                .split(':')
-                .next()
-                .unwrap_or(refspec)
-                .replace("refs/heads/", "")
-        );
-        let remote_sha = git_service
-            .read()
-            .await
-            .resolve_ref(&remote_ref_name)
-            .await
-            .unwrap_or_else(|| "0".repeat(40));
-
-        refs_lines.push(format!("{local_ref} {local_sha} {remote_ref} {remote_sha}"));
-    }
-
-    refs_lines.join("\n") + "\n"
-}
 
 #[tauri::command]
 #[specta::specta]
@@ -190,7 +140,11 @@ pub async fn push_remote(
             .flatten()
             .unwrap_or_default();
 
-        let refs_stdin = build_push_refs_stdin(&git_service, &remote_name, &refspecs).await;
+        let refs_stdin = git_service
+            .read()
+            .await
+            .build_push_refs_stdin(&remote_name, &refspecs)
+            .await;
 
         let hook_result = git_service
             .write()
@@ -261,7 +215,11 @@ pub async fn push_current_branch(
                 .unwrap_or_default();
 
             let refspecs = vec![branch.clone()];
-            let refs_stdin = build_push_refs_stdin(&git_service, &remote_name, &refspecs).await;
+            let refs_stdin = git_service
+                .read()
+                .await
+                .build_push_refs_stdin(&remote_name, &refspecs)
+                .await;
 
             let hook_result = git_service
                 .write()
